requirements.txt
geopy==2.4.1
matplotlib==3.10.3
requests==2.32.3
python-dotenv==1.1.1